rest update

+++ src/main/java/com/ClienteApiRestSnider/Controllers/ClientController.java
@@ -8,29 +8,37 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping(path = "api/clients")
+@RequestMapping(path = "api/client")
 public class ClientController {
 	@Autowired
 	private ClientService service;
 
 	@PostMapping(path = "/")
 	public ResponseEntity create(@RequestBody ClientModel model) throws Exception {
-		return new ResponseEntity<>(this.service.create(model), HttpStatus.CREATED);
+		var createdEntity = this.service.create(model);
+		if (createdEntity == null) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		return new ResponseEntity<>(createdEntity, HttpStatus.CREATED);
 	}
 
 	@PutMapping(path = "/{id}")
 	public ResponseEntity update(ClientModel model, Long id) throws Throwable, Exception {
-		return new ResponseEntity<>(this.service.update(model, id), HttpStatus.CREATED);
+		var updatedEntity = this.service.update(model, id);
+		if (updatedEntity == null) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		return new ResponseEntity<>(this.service.update(updatedEntity, id), HttpStatus.CREATED);
 	}
 
 	@DeleteMapping(path = "/{id}")
 	public ResponseEntity delete(@PathVariable Long id) throws Exception {
+		var deletedEntity = this.service.delete(id);
+		if (deletedEntity == null) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 		return new ResponseEntity<>(this.service.delete(id), HttpStatus.OK);
 	}
 
-	@GetMapping(path = "/{id}")
-	public ResponseEntity findById(Long id) throws Throwable, Exception {
-		return new ResponseEntity<>(this.service.findById(id), HttpStatus.OK);
+	@GetMapping(path = "/id/{id}")
+	public ResponseEntity findById(@PathVariable Long id) throws Throwable, Exception {
+		var foundEntity = this.service.findById(id);
+		if (foundEntity == null) return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		return new ResponseEntity<>(foundEntity, HttpStatus.OK);
 	}
 
 	@GetMapping(path = "/")
@@ -38,9 +46,11 @@ public class ClientController {
 		return new ResponseEntity<>(this.service.findAll(), HttpStatus.OK);
 	}
 
-	@GetMapping(path = "/{docNumber}")
+	@GetMapping(path = "/docnumber/{docNumber}")
 	public ResponseEntity getByDocNumber(@PathVariable String docNumber) throws Exception {
-		return new ResponseEntity<>(this.service.findByDocNumber(docNumber), HttpStatus.OK);
+		var foundEntity = this.service.findByDocNumber(docNumber);
+		if (foundEntity == null) return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		return new ResponseEntity<>(foundEntity, HttpStatus.OK);
 	}
 
 }
